sc2: Moving many client functions to apworld package; adjusted bank initialization on map launch
