+++ worlds/sc2/client.py
@@ -24,12 +24,12 @@ from typing import Any, TYPE_CHECKING, NamedTuple, Type, Sequence, Iterable
 # CommonClient import first to trigger ModuleUpdater
 from CommonClient import CommonContext, server_loop, ClientCommandProcessor, gui_enabled, get_base_parser, handle_url_arg
 from Utils import init_logging, async_start
-from .item import item_names, item_parents, race_to_item_type
+from .item import item_parents
 from .item.item_annotations import ITEM_NAME_ANNOTATIONS
-from .item.item_groups import item_name_groups, unlisted_item_name_groups, ItemGroupNames
-from . import options, VICTORY_MODULO
+from .item.item_groups import item_name_groups, unlisted_item_name_groups
+from . import options
 from .options import (
-    MissionOrder, KerriganPrimalStatus, kerrigan_unit_available, KerriganPresence, EnableMorphling, GameDifficulty,
+    MissionOrder, KerriganPrimalStatus, KerriganPresence, EnableMorphling, GameDifficulty,
     GameSpeed, GenericUpgradeItems, GenericUpgradeResearch, ColorChoice, GenericUpgradeMissions, MaxUpgradeLevel,
     LocationInclusion, ExtraLocations, MasteryLocations, SpeedrunLocations, PreventativeLocations, ChallengeLocations,
     VanillaLocations, NovaPresence,
@@ -37,53 +37,39 @@ from .options import (
     SpearOfAdunPresence, SpearOfAdunPresentInNoBuild, SpearOfAdunPassiveAbilityPresence,
     SpearOfAdunPassivesPresentInNoBuild, EnableVoidTrade, VoidTradeAgeLimit, void_trade_age_limits_ms, VoidTradeWorkers,
     DifficultyDamageModifier, MissionOrderScouting, GenericUpgradeResearchSpeedup, MercenaryHighlanders, WarCouncilNerfs,
-    is_mission_in_soa_presence,
-    upgrade_included_names,
 )
 from .mission_order.slot_data import CampaignSlotData, LayoutSlotData, MissionSlotData, MissionOrderObjectSlotData
 from .mission_order.entry_rules import SubRuleRuleData, CountMissionsRuleData, MissionEntryRules
-from .mission_tables import MissionFlag
 from .tables import NovaPresenceOptions
-from .transfer_data import normalized_unit_types, worker_units
+from .apclient.transfer_data import worker_units
 from . import SC2World
-from .apclient import user_paths
+from .apclient import banks, user_paths, game_client
 from .apclient.failable import Error
 
-
-if __name__ == "__main__":
-    init_logging("SC2Client", exception_logger="Client")
-
-sc2_logger = logging.getLogger("Starcraft2")
-
 import nest_asyncio
-from worlds._sc2common import bot
-from worlds._sc2common.bot.data import Race
-from worlds._sc2common.bot.main import run_game
-from worlds._sc2common.bot.player import Bot
-from .item.item_tables import (
-    lookup_id_to_name, ItemData,
-    ZergItemType, upgrade_bundles,
-    WEAPON_ARMOR_UPGRADE_MAX_LEVEL,
-)
 from .item import item_tables
-from .locations import SC2WOL_LOC_ID_OFFSET, LocationType, LocationFlag, SC2HOTS_LOC_ID_OFFSET, VICTORY_CACHE_OFFSET
+from .locations import (
+    SC2WOL_LOC_ID_OFFSET, LocationType, LocationFlag, SC2HOTS_LOC_ID_OFFSET, get_location_id, VICTORY_MODULO
+)
 from .mission_tables import (
     lookup_id_to_mission, SC2Campaign, MissionInfo,
     lookup_id_to_campaign, SC2Mission, campaign_mission_table, SC2Race
 )
-
 import colorama
-from NetUtils import ClientStatus, NetworkItem, JSONtoTextParser, JSONMessagePart, add_json_item, add_json_location, add_json_text, JSONTypes
+from NetUtils import (
+    NetworkItem, JSONtoTextParser, JSONMessagePart, add_json_item, add_json_location, add_json_text, JSONTypes
+)
 from MultiServer import mark_raw
-from . import banks
 
 if TYPE_CHECKING:
     from Options import Option
 
+
+sc2_logger = logging.getLogger("Starcraft2")
+
 pool = concurrent.futures.ThreadPoolExecutor(1)
 loop = asyncio.get_event_loop_policy().new_event_loop()
 nest_asyncio.apply(loop)
-MAX_BONUS: int = 28
 
 # GitHub repo where the Map/mod data is hosted for /download_data command
 DATA_REPO_OWNER = "archipelago-sc2"
@@ -121,6 +107,7 @@ def _remap_color_option(slot_data_version: int, color: int) -> int:
 class ConfigurableOptionType(enum.Enum):
     INTEGER = enum.auto()
     ENUM = enum.auto()
+
 
 class ConfigurableOptionInfo(NamedTuple):
     name: str
@@ -196,7 +183,6 @@ class StarcraftClientProcessor(ClientCommandProcessor):
                 self.output("Current difficulty: " + ["Casual", "Normal", "Hard", "Brutal"][current_difficulty])
             self.output("To change the difficulty, add the name of the difficulty after the command.")
             return False
-
 
     def _cmd_game_speed(self, game_speed: str = "") -> bool:
         """Overrides the current game speed for the world.
@@ -641,7 +627,7 @@ class SC2Context(CommonContext):
         self.difficulty_override = -1
         self.game_speed_override = -1
         self.mission_id_to_location_ids: dict[int, list[int]] = {}
-        self.last_bot: ArchipelagoBot | None = None
+        self.mission_client: game_client.MissionClient | None = None
         self.slot_data_version = 2
         self.required_tactics: int = RequiredTactics.default
         self.grant_story_tech: int = GrantStoryTech.default
@@ -1030,11 +1016,6 @@ class SC2Context(CommonContext):
 
     async def shutdown(self) -> None:
         await super(SC2Context, self).shutdown()
-        if self.last_bot:
-            self.last_bot.want_close = True
-            # If the client is not set up yet, the game is not done loading and must be force-closed
-            if not hasattr(self.last_bot, "client"):
-                bot.sc2process.kill_switch.kill_all()
         if self.sc2_process and self.sc2_process.poll() is None:
             self.sc2_process.kill()
 
@@ -1044,21 +1025,14 @@ class SC2Context(CommonContext):
 
     def play_mission(self, mission_id: int) -> bool:
         if self.missions_unlocked or is_mission_available(self, mission_id):
-            if self.sc2_process and self.sc2_process.poll() is None:
-                sc2_logger.warning("Starcraft 2 game is still running")
+            if self.mission_client and not self.mission_client.is_game_closed():
+                sc2_logger.error("Starcraft 2 game is still running")
                 return False
-            # clean up locations bank from previous map
-            banks.file_cleanup()
-            if self.slot is None:
-                sc2_logger.warning(
-                    "Launching Mission without Archipelago authentication, "
-                    "checks will not be registered to server."
-                )
-            sc2_process = starcraft_launch(self, mission_id)
-            if isinstance(sc2_process, Error):
-                sc2_logger.warning(sc2_process.message)
+            mission_client = game_client.launch_game_client(self, mission_id)
+            if isinstance(mission_client, Error):
+                sc2_logger.error(mission_client.message)
                 return False
-            self.sc2_process = sc2_process
+            self.mission_client = mission_client
             return True
         else:
             sc2_logger.info(f"{lookup_id_to_mission[mission_id].mission_name} is not currently unlocked.")
@@ -1115,18 +1089,19 @@ class SC2Context(CommonContext):
         until successful. Otherwise it will return `None` if it fails to
         acquire the lock.
         """
-        while not self.exit_event.is_set() and self.last_bot and self.last_bot.game_running:
+        while not self.exit_event.is_set() and self.mission_client and not self.mission_client.is_game_closed():
             lock = int(time.time_ns() / 1000000000) # in seconds
 
             # Make sure we're not past the waiting limit
             # SC2 needs to be notified within 10 minutes of game time (training time of the dummy units)
+            mission_time_seconds = self.mission_client.game_time_seconds()
             if self.trade_lock_start is not None:
-                if self.last_bot.time - self.trade_lock_start >= TRADE_LOCK_WAIT_LIMIT:
+                if mission_time_seconds - self.trade_lock_start >= TRADE_LOCK_WAIT_LIMIT:
                     self.trade_lock_wait = 0
                     self.trade_lock_start = None
                     return None
             elif keep_trying:
-                self.trade_lock_start = self.last_bot.time
+                self.trade_lock_start = mission_time_seconds
 
             message_uuid = str(uuid.uuid4())
             await self.send_msgs([{
@@ -1312,13 +1287,6 @@ class SC2Context(CommonContext):
         self.trade_underway = False
 
 
-
-
-class CompatItemHolder(NamedTuple):
-    name: str
-    quantity: int = 1
-
-
 async def main(arguments: Sequence[str] | None):
     multiprocessing.freeze_support()
     parser = get_base_parser()
@@ -1342,759 +1310,6 @@ async def main(arguments: Sequence[str] | None):
 
     await ctx.shutdown()
 
-# These items must be given to the player if the game is generated on older versions
-API2_TO_API3_COMPAT_ITEMS: set[CompatItemHolder] = {
-    CompatItemHolder(item_names.PHOTON_CANNON),
-    CompatItemHolder(item_names.OBSERVER),
-    CompatItemHolder(item_names.WARP_HARMONIZATION),
-    CompatItemHolder(item_names.PROGRESSIVE_PROTOSS_WEAPON_ARMOR_UPGRADE, 3)
-}
-API3_TO_API4_COMPAT_ITEMS: set[CompatItemHolder] = {
-    # War Council
-    CompatItemHolder(item_names.ZEALOT_WHIRLWIND),
-    CompatItemHolder(item_names.CENTURION_RESOURCE_EFFICIENCY),
-    CompatItemHolder(item_names.SENTINEL_RESOURCE_EFFICIENCY),
-    CompatItemHolder(item_names.STALKER_PHASE_REACTOR),
-    CompatItemHolder(item_names.DRAGOON_PHALANX_SUIT),
-    CompatItemHolder(item_names.INSTIGATOR_MODERNIZED_SERVOS),
-    CompatItemHolder(item_names.ADEPT_DISRUPTIVE_TRANSFER),
-    CompatItemHolder(item_names.SLAYER_PHASE_BLINK),
-    CompatItemHolder(item_names.AVENGER_KRYHAS_CLOAK),
-    CompatItemHolder(item_names.DARK_TEMPLAR_LESSER_SHADOW_FURY),
-    CompatItemHolder(item_names.DARK_TEMPLAR_GREATER_SHADOW_FURY),
-    CompatItemHolder(item_names.BLOOD_HUNTER_BRUTAL_EFFICIENCY),
-    CompatItemHolder(item_names.SENTRY_DOUBLE_SHIELD_RECHARGE),
-    CompatItemHolder(item_names.ENERGIZER_MOBILE_CHRONO_BEAM),
-    CompatItemHolder(item_names.HAVOC_ENDURING_SIGHT),
-    CompatItemHolder(item_names.HIGH_TEMPLAR_PLASMA_SURGE),
-    CompatItemHolder(item_names.SIGNIFIER_FEEDBACK),
-    CompatItemHolder(item_names.ASCENDANT_BREATH_OF_CREATION),
-    CompatItemHolder(item_names.DARK_ARCHON_INDOMITABLE_WILL),
-    CompatItemHolder(item_names.IMMORTAL_IMPROVED_BARRIER),
-    CompatItemHolder(item_names.VANGUARD_RAPIDFIRE_CANNON),
-    CompatItemHolder(item_names.VANGUARD_FUSION_MORTARS),
-    CompatItemHolder(item_names.ANNIHILATOR_TWILIGHT_CHASSIS),
-    CompatItemHolder(item_names.COLOSSUS_FIRE_LANCE),
-    CompatItemHolder(item_names.WRATHWALKER_AERIAL_TRACKING),
-    CompatItemHolder(item_names.REAVER_KHALAI_REPLICATORS),
-    CompatItemHolder(item_names.PHOENIX_DOUBLE_GRAVITON_BEAM),
-    CompatItemHolder(item_names.CORSAIR_NETWORK_DISRUPTION),
-    CompatItemHolder(item_names.MIRAGE_GRAVITON_BEAM),
-    CompatItemHolder(item_names.VOID_RAY_PRISMATIC_RANGE),
-    CompatItemHolder(item_names.CARRIER_REPAIR_DRONES),
-    CompatItemHolder(item_names.TEMPEST_DISINTEGRATION),
-    CompatItemHolder(item_names.ARBITER_VESSEL_OF_THE_CONCLAVE),
-    CompatItemHolder(item_names.MOTHERSHIP_TALDARIM_INTEGRATED_POWER),
-    # Other items
-    CompatItemHolder(item_names.ASCENDANT_ARCHON_MERGE),
-    CompatItemHolder(item_names.DARK_TEMPLAR_ARCHON_MERGE),
-    CompatItemHolder(item_names.SPORE_CRAWLER_BIO_BONUS),
-}
-
-def compat_item_to_network_items(compat_item: CompatItemHolder) -> list[NetworkItem]:
-    item_id = item_tables.item_table[compat_item.name].code
-    network_item = NetworkItem(item_id, 0, 0, 0)
-    return compat_item.quantity * [network_item]
-
-
-def calculate_items(ctx: SC2Context) -> dict[SC2Race, list[int]]:
-    items = ctx.items_received.copy()
-    item_list = item_tables.item_table
-    def create_network_item(item_name: str) -> NetworkItem:
-        return NetworkItem(item_list[item_name].code, 0, 0, 0)
-
-    # Items unlocked in earlier generator versions by default (Prophecy defaults, war council, rebalances)
-    if ctx.slot_data_version < 3:
-        for compat_item in API2_TO_API3_COMPAT_ITEMS:
-            items.extend(compat_item_to_network_items(compat_item))
-    if ctx.slot_data_version < 4:
-        for compat_item in API3_TO_API4_COMPAT_ITEMS:
-            items.extend(compat_item_to_network_items(compat_item))
-        received_item_ids = set(item.item for item in ctx.items_received)
-        if item_list[item_names.GHOST_RESOURCE_EFFICIENCY].code in received_item_ids:
-            items.append(create_network_item(item_names.GHOST_BARGAIN_BIN_PRICES))
-        if item_list[item_names.SPECTRE_RESOURCE_EFFICIENCY].code in received_item_ids:
-            items.append(create_network_item(item_names.SPECTRE_BARGAIN_BIN_PRICES))
-        if item_list[item_names.ROGUE_FORCES].code in received_item_ids:
-            items.append(create_network_item(item_names.UNRESTRICTED_MUTATION))
-        if item_list[item_names.SCOUT_RESOURCE_EFFICIENCY].code in received_item_ids:
-            items.append(create_network_item(item_names.SCOUT_SUPPLY_EFFICIENCY))
-        if item_list[item_names.REAVER_RESOURCE_EFFICIENCY].code in received_item_ids:
-            items.append(create_network_item(item_names.REAVER_BARGAIN_BIN_PRICES))
-
-    # API < 4 Orbital Command Count (Deprecated item)
-    orbital_command_count: int = 0
-
-    network_item: NetworkItem
-    accumulators: dict[SC2Race, list[int]] = {
-        race: [0 for element in item_type_enum_class if element.flag_word >= 0]
-        for race, item_type_enum_class in race_to_item_type.items()
-    }
-
-    # Protoss Shield grouped item specific logic
-    shields_from_ground_upgrade: int = 0
-    shields_from_air_upgrade: int = 0
-
-    for network_item in items:
-        name = lookup_id_to_name.get(network_item.item)
-        if name is None:
-            continue
-        item_data: ItemData = item_list[name]
-
-        if item_data.type.flag_word < 0:
-            continue
-
-        # exists exactly once
-        if item_data.quantity == 1 or name in item_name_groups[ItemGroupNames.UNRELEASED_ITEMS]:
-            accumulators[item_data.race][item_data.type.flag_word] |= 1 << item_data.number
-
-        # exists multiple times
-        elif item_data.quantity > 1:
-            flaggroup = item_data.type.flag_word
-
-            # Generic upgrades apply only to Weapon / Armor upgrades
-            if item_data.number >= 0:
-                accumulators[item_data.race][flaggroup] += 1 << item_data.number
-            else:
-                if name == item_names.PROGRESSIVE_PROTOSS_GROUND_UPGRADE:
-                    shields_from_ground_upgrade += 1
-                if name == item_names.PROGRESSIVE_PROTOSS_AIR_UPGRADE:
-                    shields_from_air_upgrade += 1
-                for bundled_number in get_bundle_upgrade_member_numbers(name):
-                    accumulators[item_data.race][flaggroup] += 1 << bundled_number
-
-            # Regen bio-steel nerf with API3 - undo for older games
-            if ctx.slot_data_version < 3 and name == item_names.PROGRESSIVE_REGENERATIVE_BIO_STEEL:
-                current_level = (accumulators[item_data.race][flaggroup] >> item_data.number) % 4
-                if current_level == 2:
-                    # Switch from level 2 to level 3 for compatibility
-                    accumulators[item_data.race][flaggroup] += 1 << item_data.number
-        # sum
-        # Fillers, deprecated items
-        else:
-            if name == item_names.PROGRESSIVE_ORBITAL_COMMAND:
-                orbital_command_count += 1
-            elif item_data.type == ZergItemType.Level:
-                accumulators[item_data.race][item_data.type.flag_word] += item_data.number
-            elif name == item_names.STARTING_MINERALS:
-                accumulators[item_data.race][item_data.type.flag_word] += ctx.minerals_per_item
-            elif name == item_names.STARTING_VESPENE:
-                accumulators[item_data.race][item_data.type.flag_word] += ctx.vespene_per_item
-            elif name == item_names.STARTING_SUPPLY:
-                accumulators[item_data.race][item_data.type.flag_word] += ctx.starting_supply_per_item
-            elif name == item_names.UPGRADE_RESEARCH_COST:
-                accumulators[item_data.race][item_data.type.flag_word] += ctx.research_cost_reduction_per_item
-            else:
-                accumulators[item_data.race][item_data.type.flag_word] += 1
-
-    # Fix Shields from generic upgrades by unit class (Maximum of ground/air upgrades)
-    if shields_from_ground_upgrade > 0 or shields_from_air_upgrade > 0:
-        shield_upgrade_level = max(shields_from_ground_upgrade, shields_from_air_upgrade)
-        shield_upgrade_item = item_list[item_names.PROGRESSIVE_PROTOSS_SHIELDS]
-        for _ in range(0, shield_upgrade_level):
-            accumulators[shield_upgrade_item.race][shield_upgrade_item.type.flag_word] += 1 << shield_upgrade_item.number
-
-    # Deprecated Orbital Command handling (Backwards compatibility):
-    if orbital_command_count > 0:
-        orbital_command_replacement_items: list[str] = [
-            item_names.COMMAND_CENTER_SCANNER_SWEEP,
-            item_names.COMMAND_CENTER_MULE,
-            item_names.COMMAND_CENTER_EXTRA_SUPPLIES,
-            item_names.PLANETARY_FORTRESS_ORBITAL_MODULE
-        ]
-        replacement_item_ids = [item_tables.item_table[item_name].code for item_name in orbital_command_replacement_items]
-        if sum(item_id in replacement_item_ids for item_id in items) > 0:
-            sc2_logger.warning(inspect.cleandoc("""
-                Both old Orbital Command and its replacements are present in the world. Skipping compatibility handling.
-            """))
-        else:
-            # None of replacement items are present
-            # L1: MULE and Scanner Sweep
-            scanner_sweep_data = item_tables.item_table[item_names.COMMAND_CENTER_SCANNER_SWEEP]
-            mule_data = item_tables.item_table[item_names.COMMAND_CENTER_MULE]
-            accumulators[scanner_sweep_data.race][scanner_sweep_data.type.flag_word] += 1 << scanner_sweep_data.number
-            accumulators[mule_data.race][mule_data.type.flag_word] += 1 << mule_data.number
-            if orbital_command_count >= 2:
-                # L2 MULE and Scanner Sweep usable even in Planetary Fortress Mode
-                planetary_orbital_module_data = item_tables.item_table[item_names.PLANETARY_FORTRESS_ORBITAL_MODULE]
-                accumulators[planetary_orbital_module_data.race][planetary_orbital_module_data.type.flag_word] += \
-                    1 << planetary_orbital_module_data.number
-
-    # Upgrades from completed missions
-    if ctx.generic_upgrade_missions > 0:
-        total_missions = sum(len(column) for campaign in ctx.custom_mission_order for layout in campaign.layouts for column in layout.missions)
-        num_missions = int((ctx.generic_upgrade_missions / 100) * total_missions)
-        completed = len([mission_id for mission_id in ctx.mission_id_to_location_ids if ctx.is_mission_completed(mission_id)])
-        upgrade_count = min(completed // num_missions, ctx.max_upgrade_level) if num_missions > 0 else ctx.max_upgrade_level
-        upgrade_count = min(upgrade_count, WEAPON_ARMOR_UPGRADE_MAX_LEVEL)
-
-        # Equivalent to "Progressive Weapon/Armor Upgrade" item
-        global_upgrades: set[str] = upgrade_included_names[GenericUpgradeItems.option_bundle_all]
-        for global_upgrade in global_upgrades:
-            race = item_tables.item_table[global_upgrade].race
-            upgrade_flaggroup = race_to_item_type[race]["Upgrade"].flag_word
-            for bundled_number in get_bundle_upgrade_member_numbers(global_upgrade):
-                accumulators[race][upgrade_flaggroup] += upgrade_count << bundled_number
-
-    return accumulators
-
-
-def get_bundle_upgrade_member_numbers(bundled_item: str) -> list[int]:
-    upgrade_elements: list[str] = upgrade_bundles[bundled_item]
-    if bundled_item in (item_names.PROGRESSIVE_PROTOSS_GROUND_UPGRADE, item_names.PROGRESSIVE_PROTOSS_AIR_UPGRADE):
-        # Shields are handled as a maximum of those two
-        upgrade_elements = [item_name for item_name in upgrade_elements if item_name != item_names.PROGRESSIVE_PROTOSS_SHIELDS]
-    return [item_tables.item_table[item_name].number for item_name in upgrade_elements]
-
-
-def calc_difficulty(difficulty: int):
-    if difficulty == 0:
-        return 'C'
-    elif difficulty == 1:
-        return 'N'
-    elif difficulty == 2:
-        return 'H'
-    elif difficulty == 3:
-        return 'B'
-
-    return 'X'
-
-
-def get_kerrigan_level(ctx: SC2Context, items: dict[SC2Race, list[int]], missions_beaten: int) -> int:
-    item_value = items[SC2Race.ZERG][ZergItemType.Level.flag_word]
-    mission_value = missions_beaten * ctx.kerrigan_levels_per_mission_completed
-    if ctx.kerrigan_levels_per_mission_completed_cap != -1:
-        mission_value = min(mission_value, ctx.kerrigan_levels_per_mission_completed_cap)
-    total_value = item_value + mission_value
-    if ctx.kerrigan_total_level_cap != -1:
-        total_value = min(total_value, ctx.kerrigan_total_level_cap)
-    return total_value
-
-
-def calculate_kerrigan_options(ctx: SC2Context) -> int:
-    result = 0
-
-    # Bits 0, 1
-    # Kerrigan unit available
-    if ctx.kerrigan_presence in kerrigan_unit_available:
-        result |= 1 << 0
-
-    # Bit 2
-    # Kerrigan primal status by map
-    if ctx.kerrigan_primal_status == KerriganPrimalStatus.option_vanilla:
-        result |= 1 << 2
-
-    return result
-
-def calculate_nova_presence(ctx: SC2Context, mission: SC2Mission) -> bool:
-    result = False
-    if mission.campaign == SC2Campaign.NCO:
-        if mission.race == SC2Race.TERRAN and NovaPresenceOptions.NCO_TERRAN in ctx.nova_presence:
-            result = True
-        elif mission.race == SC2Race.ZERG and NovaPresenceOptions.NCO_ZERG in ctx.nova_presence:
-            result = True
-        elif mission.race == SC2Race.PROTOSS and NovaPresenceOptions.NCO_PROTOSS in ctx.nova_presence:
-            result = True
-    if mission == SC2Mission.GHOST_OF_A_CHANCE and NovaPresenceOptions.GHOST_OF_A_CHANCE in ctx.nova_presence:
-        result = True
-    return result
-
-def calculate_story_tech(ctx: SC2Context, mission: SC2Mission) -> bool:
-    if (
-        MissionFlag.Nova in mission.flags
-        and MissionFlag.NoBuild in mission.flags
-        and ctx.nova_grant_story_tech
-    ):
-        result = GrantStoryTech.option_grant
-    else:
-        result = ctx.grant_story_tech
-    return result != 0
-
-def calculate_soa_options(ctx: SC2Context, mission: SC2Mission) -> int:
-    """
-    Pack SOA options into a single integer with bitflags.
-    0b000011 = SOA presence
-    0b000100 = SOA in no-builds
-    0b011000 = Passives presence
-    0b100000 = PAssives in no-builds
-    """
-    result = 0
-
-    # Bits 0, 1
-    # SoA Calldowns available
-    soa_presence_value = 0
-    if is_mission_in_soa_presence(ctx.spear_of_adun_presence, mission):
-        soa_presence_value = 3
-    result |= soa_presence_value << 0
-
-    # Bit 2
-    # SoA Calldowns for no-builds
-    if ctx.spear_of_adun_present_in_no_build == SpearOfAdunPresentInNoBuild.option_true:
-        result |= 1 << 2
-
-    # Bits 3,4
-    # Autocasts
-    soa_autocasts_presence_value = 0
-    if is_mission_in_soa_presence(ctx.spear_of_adun_passive_ability_presence, mission, SpearOfAdunPassiveAbilityPresence):
-        soa_autocasts_presence_value = 3
-    # Guardian Shell breaks without SoA on version 4+, but can be generated without SoA on version 3
-    if ctx.slot_data_version < 4 and MissionFlag.Protoss in mission.flags:
-        soa_autocasts_presence_value = 3
-    result |= soa_autocasts_presence_value << 3
-
-    # Bit 5
-    # Autocasts in no-builds
-    if ctx.spear_of_adun_passive_present_in_no_build == SpearOfAdunPassivesPresentInNoBuild.option_true:
-        result |= 1 << 5
-
-    return result
-
-def calculate_generic_upgrade_options(ctx: SC2Context) -> int:
-    result = 0
-
-    # Bits 0,1
-    # Research mode
-    research_mode_value = 0
-    if ctx.generic_upgrade_research == GenericUpgradeResearch.option_vanilla:
-        research_mode_value = 0
-    elif ctx.generic_upgrade_research == GenericUpgradeResearch.option_auto_in_no_build:
-        research_mode_value = 1
-    elif ctx.generic_upgrade_research == GenericUpgradeResearch.option_auto_in_build:
-        research_mode_value = 2
-    elif ctx.generic_upgrade_research == GenericUpgradeResearch.option_always_auto:
-        research_mode_value = 3
-    result |= research_mode_value << 0
-
-    # Bit 2
-    # Speedup
-    if ctx.generic_upgrade_research_speedup == GenericUpgradeResearchSpeedup.option_true:
-        result |= 1 << 2
-
-    return result
-
-def calculate_trade_options(ctx: SC2Context) -> int:
-    result = 0
-
-    # Bit 0
-    # Trade enabled
-    if ctx.trade_enabled:
-        result |= 1 << 0
-
-    # Bit 1
-    # Workers allowed
-    if ctx.trade_workers_allowed == VoidTradeWorkers.option_true:
-        result |= 1 << 1
-
-    return result
-
-def kerrigan_primal(ctx: SC2Context, kerrigan_level: int) -> bool:
-    if ctx.kerrigan_primal_status == KerriganPrimalStatus.option_always_zerg:
-        return True
-    elif ctx.kerrigan_primal_status == KerriganPrimalStatus.option_always_human:
-        return False
-    elif ctx.kerrigan_primal_status == KerriganPrimalStatus.option_level_35:
-        return kerrigan_level >= 35
-    elif ctx.kerrigan_primal_status == KerriganPrimalStatus.option_half_completion:
-        total_missions = len(ctx.mission_id_to_location_ids)
-        completed = sum(ctx.is_mission_completed(mission_id)
-                         for mission_id in ctx.mission_id_to_location_ids)
-        return completed >= (total_missions / 2)
-    elif ctx.kerrigan_primal_status == KerriganPrimalStatus.option_item:
-        codes = [item.item for item in ctx.items_received]
-        return item_tables.item_table[item_names.KERRIGAN_PRIMAL_FORM].code in codes
-    return False
-
-
-def get_mission_variant(mission_id: int) -> int:
-    mission_flags = lookup_id_to_mission[mission_id].flags
-    if MissionFlag.RaceSwap not in mission_flags:
-        return 0
-    if MissionFlag.Terran in mission_flags:
-        return 1
-    elif MissionFlag.Zerg in mission_flags:
-        return 2
-    elif MissionFlag.Protoss in mission_flags:
-        return 3
-    return 0
-
-
-def get_item_flag_word(item_name: str) -> int:
-    return item_tables.item_table[item_name].type.flag_word
-
-
-def starcraft_launch(ctx: SC2Context, mission_id: int) -> subprocess.Popen | Error[str]:
-    sc2_logger.info(f"Launching {lookup_id_to_mission[mission_id].mission_name}.")
-    sc2_install_dir = user_paths.get_sc2_install_dir()
-    if isinstance(sc2_install_dir, Error):
-        return sc2_install_dir
-    sc2_switcher_exe = os.path.join(sc2_install_dir, "Support64", "SC2Switcher_x64.exe")
-    difficulty = ctx.difficulty if ctx.difficulty_override < 0 else ctx.difficulty_override
-    # Note(mm): SC2Switcher accepts difficulty on a scale of 1~4, the Archipelago option is 0~3, though
-    difficulty = min(4, difficulty + 1)
-    speed = ctx.game_speed if ctx.game_speed_override < 0 else ctx.game_speed_override
-    TRIGDEBUG = False
-    mission = lookup_id_to_mission[mission_id]
-    map_path = os.path.join("ArchipelagoCampaign", mission.campaign.folder, mission.map_file + ".SC2Map")
-    command_line = [
-        sc2_switcher_exe,
-        "-run", map_path,
-        # Note(mm): displaymode 2 is available for true fullscreen
-        "-displaymode", ("0" if SC2World.settings.game_windowed_mode else "1"),
-        "-difficulty", str(difficulty),
-    ]
-    if TRIGDEBUG:
-        command_line.append("-trigdebug")
-    if speed != GameSpeed.option_default:
-        # Note(mm): SC2Switcher accepts speed on a scale of 0~5, the Archipelago option 0 is "default" though
-        speed = max(0, speed - 1)
-        command_line.extend(("-speed", str(speed)))
-    if sys.platform == "win32":
-        with DllDirectory(None):
-            sc2_logger.debug(command_line)
-            return subprocess.Popen(command_line)
-
-    # Linux
-    wine = user_paths.get_wine_path()
-    if isinstance(wine, Error):
-        return wine
-    wine_prefix = user_paths.get_wine_prefix()
-    if isinstance(wine_prefix, Error):
-        return wine_prefix
-    command_line = [wine] + command_line
-    sc2_logger.debug(f"WINEPREFIX: {wine_prefix}")
-    sc2_logger.debug(command_line)
-    return subprocess.Popen(
-        command_line,
-        env=os.environ | {"WINEPREFIX": wine_prefix},
-        stderr=subprocess.DEVNULL,
-    )
-
-
-class ArchipelagoBot(bot.bot_ai.BotAI):
-    __slots__ = [
-        'game_running',
-        'mission_completed',
-        'boni',
-        'setup_done',
-        'ctx',
-        'mission_id',
-        'want_close',
-        'can_read_game',
-        'last_received_update',
-        'last_trade_cargo',
-        'last_supply_used'
-    ]
-    ctx: SC2Context
-    # defined in bot_ai_internal.py; seems to be mis-annotated as a float and later re-annotated as an int
-    supply_used: int
-
-    def __init__(self, ctx: SC2Context, mission_id: int):
-        self.game_running = False
-        self.mission_completed = False
-        self.want_close = False
-        self.can_read_game = False
-        self.last_received_update: int = 0
-        self.last_trade_cargo: set = set()
-        self.last_supply_used: int = 0
-        self.trade_reply_cooldown: int = 0
-        self.setup_done = False
-        self.ctx = ctx
-        self.ctx.last_bot = self
-        self.mission_id = mission_id
-        self.boni = [False for _ in range(MAX_BONUS)]
-
-        super(ArchipelagoBot, self).__init__()
-
-    async def on_step(self, iteration: int) -> None:
-        if self.want_close:
-            self.want_close = False
-            await self._client.leave()
-            return
-        game_state = 0
-        if not self.setup_done:
-            self.setup_done = True
-            mission = lookup_id_to_mission[self.mission_id]
-            start_items = calculate_items(self.ctx)
-            missions_beaten = self.missions_beaten_count()
-            kerrigan_level = get_kerrigan_level(self.ctx, start_items, missions_beaten)
-            kerrigan_options = calculate_kerrigan_options(self.ctx)
-            nova_presence = calculate_nova_presence(self.ctx, mission)
-            grant_story_tech = calculate_story_tech(self.ctx, mission)
-            soa_options = calculate_soa_options(self.ctx, mission)
-            generic_upgrade_options = calculate_generic_upgrade_options(self.ctx)
-            trade_options = calculate_trade_options(self.ctx)
-            mission_variant = get_mission_variant(self.mission_id)  # 0/1/2/3 for unchanged/Terran/Zerg/Protoss
-            uncollected_objectives: list[int] = self.get_uncollected_objectives()
-            if self.ctx.difficulty_override >= 0:
-                difficulty = calc_difficulty(self.ctx.difficulty_override)
-            else:
-                difficulty = calc_difficulty(self.ctx.difficulty)
-            if self.ctx.game_speed_override >= 0:
-                game_speed = self.ctx.game_speed_override
-            else:
-                game_speed = self.ctx.game_speed
-
-            error = banks.send_options(
-                f" {difficulty}"
-                f" {generic_upgrade_options}"
-                f" {self.ctx.all_in_choice}"
-                f" {game_speed}"
-                f" {self.ctx.disable_forced_camera}"
-                f" {self.ctx.skip_cutscenes}"
-                f" {kerrigan_options}"
-                f" {grant_story_tech}"
-                f" {self.ctx.take_over_ai_allies}"
-                f" {soa_options}"
-                f" {self.ctx.mission_order}"
-                f" {int(nova_presence)}"
-                f" {self.ctx.grant_story_levels}"
-                f" {self.ctx.enable_morphling}"
-                f" {mission_variant}"
-                f" {trade_options}"
-                f" {self.ctx.difficulty_damage_modifier}"
-                f" {self.ctx.mercenary_highlanders}" # TODO: Possibly rework into unit options
-                f" {self.ctx.war_council_nerfs}"
-            )
-            if isinstance(error, Error):
-                sc2_logger.error(error.message)
-                return
-            self.update_tech(start_items, kerrigan_level)
-            objectives = ""
-            if uncollected_objectives:
-                objectives = " ".join(f"{str(objective)}" for objective in uncollected_objectives)
-            error = banks.send_core_options(
-                self.get_resources(start_items),
-                self.get_colors(),
-                objectives,
-                "1"
-            )
-            if isinstance(error, Error):
-                sc2_logger.error(error.message)
-                return
-            self.last_received_update = len(self.ctx.items_received)
-        else:
-            error = banks.send_ap_messages_from_queue(self.ctx.announcements)
-            if isinstance(error, Error):
-                sc2_logger.error(error.message)
-                return
-            trade_send_string = self.get_trade_units_sent()
-            # Message format:
-            # <unit1> <unit2> <unit3>...
-            if len(trade_send_string) > 0:
-                units_to_send: list[str] = []
-                non_ap_units: set[str] = set()
-                for unit_id in trade_send_string.split():
-                    if unit_id.startswith("AP_"):
-                        units_to_send.append(normalized_unit_types.get(unit_id, unit_id))
-                    else:
-                        non_ap_units.add(unit_id)
-                if len(non_ap_units) > 0:
-                    sc2_logger.info(f"Void Trade tried to send non-AP units: {', '.join(non_ap_units)}")
-                    banks.send_trade_received("FailNonAPUnits")
-                    self.ctx.trade_underway = True
-                else:
-                    self.ctx.trade_response = None
-                    self.ctx.trade_underway = True
-                    async_start(self.ctx.trade_send(units_to_send))
-
-            trade_receive_string = self.get_trade_receive_request()
-            # Message format:
-            # <count>
-            if len(trade_receive_string) > 0:
-                if int(trade_receive_string) == 1:
-                    self.ctx.trade_underway = True
-                    self.ctx.trade_response = None
-                    async_start(self.ctx.trade_receive(1))
-                    # TODO: handle supply, self.ctx.trade_response = "?TradeFail Void Trade rejected: Not enough supply."
-                elif int(trade_receive_string) == 5:
-                    self.ctx.trade_underway = True
-                    self.ctx.trade_response = None
-                    async_start(self.ctx.trade_receive(5))
-
-            game_state = self.get_locations()
-
-            if game_state & 1:
-                self.can_read_game = True
-
-            if iteration == 160 and not game_state & 1:
-                error = banks.send_ap_message([
-                    "Warning: Archipelago unable to connect or has lost connection to "
-                    "Starcraft 2 (This is likely a map issue)"
-                ])
-                if isinstance(error, Error):
-                    sc2_logger.error(error.message)
-                    return
-
-            if banks.update_prompt():
-                self.last_received_update = 0
-
-            if self.last_received_update < len(self.ctx.items_received):
-                current_items = calculate_items(self.ctx)
-                missions_beaten = self.missions_beaten_count()
-                kerrigan_level = get_kerrigan_level(self.ctx, current_items, missions_beaten)
-                if (error := self.update_core_options(current_items)):
-                    sc2_logger.error(error.message)
-                    return
-                if (error := self.update_tech(current_items, kerrigan_level)):
-                    sc2_logger.error(error.message)
-                    return
-                self.last_received_update = len(self.ctx.items_received)
-
-            if game_state & 1:
-                if not self.game_running:
-                    error = banks.send_ap_message(["Archipelago Connected"])
-                    if isinstance(error, Error):
-                        sc2_logger.error(error.message)
-                        return
-                    self.game_running = True
-
-                if self.can_read_game:
-                    if game_state & (1 << 1) and not self.mission_completed:
-                        victory_locations = [get_location_id(self.mission_id, 0)]
-                        send_victory = (
-                            self.mission_id in self.ctx.final_mission_ids and
-                            len(self.ctx.final_locations) == len(self.ctx.checked_locations.union(victory_locations).intersection(self.ctx.final_locations))
-                        )
-
-                        # Old slots don't have locations on goal
-                        if not send_victory or self.ctx.slot_data_version >= 4:
-                            sc2_logger.info("Mission Completed")
-                            location_ids = self.ctx.mission_id_to_location_ids[self.mission_id]
-                            victory_locations += sorted([
-                                get_location_id(self.mission_id, location_id)
-                                for location_id in location_ids
-                                if (location_id % VICTORY_MODULO) >= VICTORY_CACHE_OFFSET
-                            ])
-                            await self.ctx.send_msgs(
-                                [{"cmd": 'LocationChecks',
-                                    "locations": victory_locations}])
-                            self.mission_completed = True
-
-                        if send_victory:
-                            print("Game Complete")
-                            await self.ctx.send_msgs([{"cmd": 'StatusUpdate', "status": ClientStatus.CLIENT_GOAL}])
-                            self.mission_completed = True
-                            self.ctx.finished_game = True
-
-                    for x, completed in enumerate(self.boni):
-                        if not completed and game_state & (1 << (x + 2)):
-                            await self.ctx.send_msgs(
-                                [{"cmd": 'LocationChecks',
-                                  "locations": [get_location_id(self.mission_id, x + 1)]}])
-                            self.boni[x] = True
-
-                    # Send Void Trade results
-                    if self.ctx.trade_response is not None and self.trade_reply_cooldown == 0:
-                        error = banks.send_ap_message([self.ctx.trade_response])
-                        if isinstance(error, Error):
-                            sc2_logger.error(error.message)
-                            return
-                        # Wait an arbitrary amount of frames before trying again
-                        self.trade_reply_cooldown = 60
-                else:
-                    error = banks.send_ap_message(["LostConnection - Lost connection to game."])
-                    if isinstance(error, Error):
-                        sc2_logger.error(error.message)
-                        return
-
-    def get_locations(self) -> int:
-        result = banks.read_locations()
-        if isinstance(result, Error):
-            return 0
-        if (result.strip()):  # may be "" or " "
-            return int(result)
-        return 0
-
-    def get_trade_units_sent(self) -> str:
-        result = banks.read_trade_units()
-        if isinstance(result, Error):
-            return ""
-        return result
-
-    def get_trade_receive_request(self) -> str:
-        result = banks.read_trade_receive_request()
-        if isinstance(result, Error):
-            return ""
-        return result
-
-    def get_uncollected_objectives(self) -> list[int]:
-        result = [
-            location % VICTORY_MODULO
-            for location in self.ctx.uncollected_locations_in_mission(lookup_id_to_mission[self.mission_id])
-            if (location % VICTORY_MODULO) < VICTORY_CACHE_OFFSET
-        ]
-        return result
-
-    def missions_beaten_count(self) -> int:
-        return len([location for location in self.ctx.checked_locations if location % VICTORY_MODULO == 0])
-
-    def get_colors(self) -> str:
-        self.ctx.pending_color_update = False
-        result = [
-            str(self.ctx.player_color_raynor),
-            str(self.ctx.player_color_zerg),
-            str(self.ctx.player_color_zerg_primal),
-            str(self.ctx.player_color_protoss),
-            str(self.ctx.player_color_nova),
-        ]
-        return ' '.join(result)
-
-    def get_resources(self, current_items: dict[SC2Race, list[int]]) -> str:
-        DEFAULT_MAX_SUPPLY = 200
-        max_supply_amount = max(
-            DEFAULT_MAX_SUPPLY
-            + (
-                current_items[SC2Race.ANY][get_item_flag_word(item_names.MAX_SUPPLY)]
-                * self.ctx.maximum_supply_per_item
-            )
-            - (
-                current_items[SC2Race.ANY][get_item_flag_word(item_names.REDUCED_MAX_SUPPLY)]
-                * self.ctx.maximum_supply_reduction_per_item
-            ),
-            self.ctx.lowest_maximum_supply,
-        )
-        return ("{} {} {} {}".format(
-            current_items[SC2Race.ANY][get_item_flag_word(item_names.STARTING_MINERALS)],
-            current_items[SC2Race.ANY][get_item_flag_word(item_names.STARTING_VESPENE)],
-            current_items[SC2Race.ANY][get_item_flag_word(item_names.STARTING_SUPPLY)],
-            max_supply_amount - DEFAULT_MAX_SUPPLY,
-        ))
-
-    def get_terran_tech(self, current_items: dict[SC2Race, list[int]]) -> str:
-        terran_items = current_items[SC2Race.TERRAN]
-        return (" ".join(map(str, terran_items)))
-
-    def get_zerg_tech(self, current_items: dict[SC2Race, list[int]], kerrigan_level: int) -> str:
-        zerg_items = current_items[SC2Race.ZERG]
-        zerg_items = [value for index, value in enumerate(zerg_items) if index not in [ZergItemType.Level.flag_word, ZergItemType.Primal_Form.flag_word]]
-        kerrigan_primal_by_items = kerrigan_primal(self.ctx, kerrigan_level)
-        kerrigan_primal_bot_value = 1 if kerrigan_primal_by_items else 0
-        return(f"{kerrigan_level} {kerrigan_primal_bot_value} " + ' '.join(map(str, zerg_items)))
-
-    def get_protoss_tech(self, current_items: dict[SC2Race, list[int]]) -> str:
-        protoss_items = current_items[SC2Race.PROTOSS]
-        return (" ".join(map(str, protoss_items)))
-
-    def get_misc_tech(self, current_items: dict[SC2Race, list[int]]) -> str:
-        return ("{} {} {}".format(
-            current_items[SC2Race.ANY][get_item_flag_word(item_names.BUILDING_CONSTRUCTION_SPEED)],
-            current_items[SC2Race.ANY][get_item_flag_word(item_names.UPGRADE_RESEARCH_SPEED)],
-            current_items[SC2Race.ANY][get_item_flag_word(item_names.UPGRADE_RESEARCH_COST)],
-        ))
-
-    def update_tech(self, current_items: dict[SC2Race, list[int]], kerrigan_level: int) -> None | Error[str]:
-        return banks.send_items(
-            self.get_terran_tech(current_items),
-            self.get_zerg_tech(current_items, kerrigan_level),
-            self.get_protoss_tech(current_items),
-            self.get_misc_tech(current_items)
-        )
-
-    def update_core_options(self, current_items: dict[SC2Race, list[int]]) -> None | Error[str]:
-        return banks.send_core_options(
-            self.get_resources(current_items),
-            self.get_colors()
-        )
 
 def calc_unfinished_nodes(
     ctx: SC2Context
@@ -2149,11 +1364,11 @@ def calc_available_nodes(ctx: SC2Context) -> tuple[list[int], dict[int, list[int
         accessible_objects_to_add: list[MissionOrderObjectSlotData] = []
         for mission_order_object in candidate_accessible_objects:
             if (
-                    mission_order_object.entry_rule.is_accessible(beaten_accessible_missions, received_items)
-                    and all([
-                        parent_object.entry_rule.is_accessible(beaten_accessible_missions, received_items)
-                        for parent_object in parent_objects[mission_order_objects.index(mission_order_object)]
-                    ])
+                mission_order_object.entry_rule.is_accessible(beaten_accessible_missions, received_items)
+                and all([
+                    parent_object.entry_rule.is_accessible(beaten_accessible_missions, received_items)
+                    for parent_object in parent_objects[mission_order_objects.index(mission_order_object)]
+                ])
             ):
                 accessible_objects_to_add.append(mission_order_object)
         if len(accessible_objects_to_add) > 0:
@@ -2263,46 +1478,6 @@ def is_mod_installed_correctly() -> bool:
     else:
         sc2_logger.debug("All map/mod files are properly installed.")
         return True
-
-
-class DllDirectory:
-    """
-    Context manager that sets the DLL search path on Windows and restores it on exit.
-    Setting the new path to `None` restores the default search order.
-    See https://docs.microsoft.com/en-us/windows/win32/api/winbase/nf-winbase-setdlldirectoryw
-    """
-    # Credit to Black Sliver for this code.
-    _old: str | None = None
-    _new: str | None = None
-
-    def __init__(self, new: str | None):
-        self._new = new
-
-    def __enter__(self):
-        old = self.get()
-        if self.set(self._new):
-            self._old = old
-
-    def __exit__(self, *args):
-        if self._old is not None:
-            self.set(self._old)
-
-    @staticmethod
-    def get() -> str | None:
-        if sys.platform == "win32":
-            n = ctypes.windll.kernel32.GetDllDirectoryW(0, None)
-            buf = ctypes.create_unicode_buffer(n)
-            ctypes.windll.kernel32.GetDllDirectoryW(n, buf)
-            return buf.value
-        # NOTE: other OS may support os.environ["LD_LIBRARY_PATH"], but this fix is windows-specific
-        return None
-
-    @staticmethod
-    def set(s: str | None) -> bool:
-        if sys.platform == "win32":
-            return ctypes.windll.kernel32.SetDllDirectoryW(s) != 0
-        # NOTE: other OS may support os.environ["LD_LIBRARY_PATH"], but this fix is windows-specific
-        return False
 
 
 def download_latest_release_zip(
@@ -2438,15 +1613,6 @@ def is_mod_update_available(owner: str, repo: str, api_version: str, metadata: s
         r1.close()
 
 
-def get_location_offset(mission_id: int) -> int:
-    return SC2WOL_LOC_ID_OFFSET if mission_id <= SC2Mission.ALL_IN.id \
-        else (SC2HOTS_LOC_ID_OFFSET - SC2Mission.ALL_IN.id * VICTORY_MODULO)
-
-
-def get_location_id(mission_id: int, objective_id: int) -> int:
-    return get_location_offset(mission_id) + mission_id * VICTORY_MODULO + objective_id
-
-
 _has_forced_save = False
 def force_settings_save_on_close() -> None:
     """
@@ -2463,5 +1629,6 @@ def force_settings_save_on_close() -> None:
 
 def launch(*args: str):
     colorama.just_fix_windows_console()
+    init_logging("SC2Client", exception_logger="Client")
     asyncio.run(main(args))
     colorama.deinit()
